# [Xamarin.Android.Build.Tasks] improvements to Set*Writeable