--- a/Documentation/release-notes/4407.md
+++ b/Documentation/release-notes/4407.md
@@ -1,0 +1,5 @@
+### Build and deployment performance
+
+  * [GitHub PR 4407](https://github.com/xamarin/xamarin-android/pull/4407):
+    Improved code for removing "readonly" attributes from files and directories.
+    This saved ~90ms to an initial build of a Xamarin.Forms template project.