--- a/src/Xamarin.Android.Build.Tasks/Tasks/ResolveLibraryProjectImports.cs
+++ b/src/Xamarin.Android.Build.Tasks/Tasks/ResolveLibraryProjectImports.cs
@@ -101,11 +101,11 @@ namespace Xamarin.Android.Tasks
 				.ToArray ();
 
 			foreach (var directory in ResolvedResourceDirectories) {
-				MonoAndroidHelper.SetDirectoryWriteable (directory.ItemSpec);
+				Files.SetDirectoryWriteable (directory.ItemSpec);
 			}
 
 			foreach (var directory in ResolvedAssetDirectories) {
-				MonoAndroidHelper.SetDirectoryWriteable (directory.ItemSpec);
+				Files.SetDirectoryWriteable (directory.ItemSpec);
 			}
 
 			if (!string.IsNullOrEmpty (CacheFile)) {
@@ -149,7 +149,7 @@ namespace Xamarin.Android.Tasks
 			// lets "upgrade" the old directory.
 			string oldPath = Path.GetFullPath (Path.Combine (OutputImportDirectory, "..", "__library_projects__"));
 			if (!OutputImportDirectory.Contains ("__library_projects__") && Directory.Exists (oldPath)) {
-				MonoAndroidHelper.SetDirectoryWriteable (Path.Combine (oldPath, ".."));
+				Files.SetDirectoryWriteable (Path.Combine (oldPath, ".."));
 				Directory.Delete (oldPath, recursive: true);
 			}
 			var outdir = Path.GetFullPath (OutputImportDirectory);
@@ -186,7 +186,7 @@ namespace Xamarin.Android.Tasks
 
 				// Skip already-extracted resources.
 				bool updated = false;
-				string assemblyHash = MonoAndroidHelper.HashFile (assemblyPath);
+				string assemblyHash = Files.HashFile (assemblyPath);
 				string stamp = Path.Combine (outdir, assemblyIdentName + ".stamp");
 				string stampHash = File.Exists (stamp) ? File.ReadAllText (stamp) : null;
 				if (assemblyHash == stampHash) {
@@ -228,7 +228,7 @@ namespace Xamarin.Android.Tasks
 						if (name.StartsWith ("__AndroidEnvironment__", StringComparison.OrdinalIgnoreCase)) {
 							var outFile = Path.Combine (outDirForDll, name);
 							using (var stream = pe.GetEmbeddedResourceStream (resource)) {
-								updated |= MonoAndroidHelper.CopyIfStreamChanged (stream, outFile);
+								updated |= Files.CopyIfStreamChanged (stream, outFile);
 							}
 							resolvedEnvironments.Add (new TaskItem (Path.GetFullPath (outFile), new Dictionary<string, string> {
 								{ OriginalFile, assemblyPath }
@@ -238,7 +238,7 @@ namespace Xamarin.Android.Tasks
 						else if (name.EndsWith (".jar", StringComparison.InvariantCultureIgnoreCase)) {
 							using (var stream = pe.GetEmbeddedResourceStream (resource)) {
 								AddJar (jars, importsDir, name, assemblyPath);
-								updated |= MonoAndroidHelper.CopyIfStreamChanged (stream, Path.Combine (importsDir, name));
+								updated |= Files.CopyIfStreamChanged (stream, Path.Combine (importsDir, name));
 							}
 						}
 						// embedded native libraries
@@ -345,7 +345,7 @@ namespace Xamarin.Android.Tasks
 				string assetsDir = Path.Combine (importsDir, "assets");
 
 				bool updated = false;
-				string aarHash = MonoAndroidHelper.HashFile (aarFile.ItemSpec);
+				string aarHash = Files.HashFile (aarFile.ItemSpec);
 				string stamp = Path.Combine (outdir, aarIdentityName + ".stamp");
 				string stampHash = File.Exists (stamp) ? File.ReadAllText (stamp) : null;
 				var aarFullPath = Path.GetFullPath (aarFile.ItemSpec);