Context: https://github.com/jonathanpeppers/Benchmarks/tree/setwriteable

`MonoAndroidHelper` has a couple methods that are called a lot during
a build:

1. `public static void SetDirectoryWriteable (string directory)`
2. `public static void SetWriteable (string source)`

Both have some room for performance improvements.

## SetWriteable ##

I changed usage of `FileInfo` to just use the static `File` methods.
We can do this for files, but not directories. There is no
`Directory.GetAttributes` method in the BCL.

We had a `File.Exists` check, which all callers were already doing. I
changed this to use `try-catch` around `File.GetAttributes`.

Benchmarking the difference:

    BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18362
    Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
    [Host]     : .NET Framework 4.8 (4.8.4121.0), X86 LegacyJIT

    |     Method |      Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
    |----------- |----------:|---------:|---------:|------:|------:|------:|----------:|
    | Writeable2 |  36.20 us | 0.531 us | 0.443 us |     - |     - |     - |         - |
    | Writeable1 |  61.27 us | 0.826 us | 0.732 us |     - |     - |     - |         - |
    |  Readonly2 | 100.97 us | 1.274 us | 1.129 us |     - |     - |     - |         - |
    |  Readonly1 | 126.21 us | 1.441 us | 1.348 us |     - |     - |     - |         - |

## SetDirectoryWriteable ##

The main improvement is to *not* call `Directory.EnumerateDirectories`
and `Directory.EnumerateFiles` independently. Use
`DirectoryInfo.EnumerateFileSystemInfos` instead. This way we aren't
enumerating the file system twice.

It also doesn't need to call `SetWriteable` for files.

Benchmarking the difference, it is an order of magnitude better now:

    BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18362
    Intel Core i9-9900K CPU 3.60GHz (Coffee Lake), 1 CPU, 16 logical and 8 physical cores
    [Host]     : .NET Framework 4.8 (4.8.4121.0), X86 LegacyJIT

    |     Method |      Mean |    Error |   StdDev | Gen 0 | Gen 1 | Gen 2 | Allocated |
    |----------- |----------:|---------:|---------:|------:|------:|------:|----------:|
    | Writeable2 | 158.51 us | 1.870 us | 1.749 us |     - |     - |     - |         - |
    |  Readonly2 | 159.54 us | 2.172 us | 2.032 us |     - |     - |     - |         - |
    | Writeable1 | 758.41 us | 5.558 us | 5.199 us |     - |     - |     - |   16384 B |
    |  Readonly1 | 763.97 us | 8.778 us | 8.211 us |     - |     - |     - |   16384 B |

## Overall results ##

On an initial build of the Xamarin.Forms integration project in this
repo:

    Before:
     461 ms  GenerateJavaStubs                          1 calls
    1061 ms  ResolveLibraryProjectImports               1 call
    After:
     448 ms  GenerateJavaStubs                          1 calls
     993 ms  ResolveLibraryProjectImports               1 calls

These two targets would call into `Set*Writeable` a lot because they
write lots of files.

This saves ~90ms on an initial build. Incremental builds will have a
smaller improvement.