@jonpryor 2020-03-17
In the `SetWritable` > Benchmarking the difference section, I don't know what the benchmarks *are*.  What is `Writeable` vs. `Writeable2` vs. `Readonly` vs. `Readonly2` regarding *files* vs. *directories*?